Updating code to better integrate DXF output
